iii,remove css styling from the templates and instead assign classes or ids so that the styles can be easily customisable;    - Done
iv, The TOC box view
    - should show the section heading;
         - Done
    - should also give a style to the currently selected page so that
it can be customised different from the rest;
         - Done. New style added and font weight changed to bold

v, When we create referencemanual pages - the file name does not save
as the title given to the page and so if we want a friendly file name
we have to edit that manually - please change that to automatically
store the title of the page as the file name.
    - Done

vi. Make the body field in the ReferenceManualPage optional
    - Done