[tests] Providing a workaround to issue that inline registrations for DC adapters getting dropped between test suites (it seems due to interference python loaded code and interference with placelesssetup).
Wrapping workaround in core.testing.refresh_dc_registrations(). Providing sampe tests that make use of it in core/workflows/question.txt and ui/forms/readme.txt.

Some test code cleanup; remove unused imports, re-ordering of zcml to more logically follow  "dependency order".

A related complication of is that the module in question here, bungenci.core.dc, is in bungeni.core, but it is really a ui concept, and should probably be in bungeni.ui.