[config] Adding "bungeni.CommitteeMember" as role that participates in form configuration (descriptors).
This fixes problem with of having certain fields (precisely all those fields in a descriptor that are localizable in "add" mode) not being displayed in the form for a new item... specifically, when creating an AgendaItem in the context of a Committee, the Clerk was seeing fields Acronym, Descrition and Language, but the MP was NOT seeing these (and, Language being required, resulting in that MP cannot create AgendaItems).

Related issue: should add/auto-extract other "member" roles.