[declarative] Splitting up UI setup from feature configuration into setup and apply, as apply (execution of the prepared zcml code) requires that interfaces and related constructs had been already generated and are defined (are done by the type being "catalysed").
Reworked code in WorkspaceDataTableFormatter.get_status to avoid IWorkflow adaptation failure on newly addded custom types; other comments.

Improved error message for when when a permission grant to a role is mixed globally and locally (as promised to Pedro).

Improved error message for when a custom type is added but no ui descriptor specified.