[containers] Re-instating the container properties (delegations, and one for each doc type) on User that for some reason disappeared in r10478 (probably because of auto-regeneration of default ui.xml, that at this rev did not support roundtripping container info, and the changes to the checked in ui.xml were not reviewed): - re-instating the delegations container property on the domain User class. - refactoring the container propery per doc type from being on User to being on Member.
Bringing back the idea of indirect relations to define a container (that was dropped in r10482) but evolving it to support that the target fk name (on parent) may be different than the source fk name (on the child/contained items).

Adding container.@indirect_key attribute to the descriptor XML schema.

Updated auto-generated workflow permissions.zcml (not included in some recent changeset triggering the changes).