[draft refactor] Cleaning out after removal of all "working_draft" states: moving "user_is_context_owner" condition to simply be a utility; dropping "user_is_not_context_owner" condition altogether; removing workflow transition conditions that were using these (all unnecessary, controlled by straightforward roles).
Relegating "context_is_public" from being a workflow condition (but being used by other conditions) as (a) it is not being used directly as a condition and (b) the logic is wrong (added revisit comment).

Removing the couple remaining "working_draft" states that had been left in as obsolete="true", for data backwards compatibility.

Misc other minor corrections and cleanup.