[workflow] Adding support for an optional single anonymous in-place facet within workflow state elements, that may be specified instead of a facet ref to a facet in the same workflow. Makes the case of having a state-dedicated facet more convenient. Key aspects of anonymous facets to note: - only one anonymous facet per state is allowed - an anonymous state facet is NOT available for re-use by another state or from another workflow
Reworking all facet names prefixed with "state_" as anonymous facets.

Correcting recent failure in the determination of default facet for a feature in a state (for when that state does not specify a facet for the feature, and the associated workflows specifies a default facet); now default facets should be determined correctly.

Reworking permissions for the "draft" and all "add_*" facets of attachment/event, to ensure that once the user has added the item, s/he can always edit/delete while the item remaining private i.e. no-one else may edit/delete it (even others who may add this kind of item).