[conditions, event response] Reworking the condition response_allow_submit(question) for new event-based response scenario, using the following newly added generic utilities as building blocks:     child(context, type_key)     in_state(context, *state_ids)
Moving to bungeni_custom the implementation of the following question-specific conditions:
    is_written_response
    is_oral_response
    response_allow_submit
Re-adding these conditions to the assembly_question and senate_question workflows.

Moving transition sub-check, of whether we are in edit mode from spurious use () to centrally, checked systematically (so no transitions should be available in the UI when UI is in edit mode).