r10928 - fri 19-apr-2013 [custom archetype usage example] Adding a custom base type "question" and making other logical question types, in this case "assembly_question" and "senate_queston" be based on this as archetype.
Currently requires that a descriptor file be provided, but soon this should not be needed; a related issue is whether to allow "inheritance" of the descriptor/workflow configs of such an intermediate custom aerchetype.

Correcting deletion of what shoyuld be unnecessary setting of change.seq in audit.py, in r10927, that is however still needed.