[office member permissions] Adjusting permissions within generic group workflow (upolitical_group, office, government, ministry types) to make members & title_types visible to all when public, and to allow Clerk to manage membership.
Moving registration of additems add and contextactions edit/delete menu items, for group and group_member to ui.interfaces.IWorkspaceOrAdminSectionLayer (instead of restricted to ui.interfaces.IAdminSectionLayer as previous) to enable access to these by the clerk.

Corrects the issue, logged in as clerk, it is now possible to view and add members as well as to view title_types (not add/manage them however, as enabling the option gives error, so this is still admin only).